Move collection post tests to separate file
